[revision-2] relazed constraints
