+++ src/main/java/it/gov/pagopa/bizpmingestion/entity/pm/PPWallet.java
@@ -40,8 +40,7 @@ public class PPWallet {
     @OneToMany(targetEntity = PPPayPal.class, fetch = FetchType.EAGER, mappedBy = "ppWallet", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<PPPayPal> ppPayPal = new ArrayList<>();
 
-    @OneToOne(targetEntity = PPPsp.class, fetch = FetchType.EAGER)
-    @JoinColumn(name = "FK_PSP", referencedColumnName = "ID", insertable = false, updatable = false)
-    private PPPsp ppPsp;
+    @Column(name = "FK_PSP")
+    private Long ppPsp;
 
 }
